Przeniesienie metod z services do repository
Metoda wykonujaca procedure dodaj czytelnika
